Update master and single capacity unit
Currently, master and single instances specify that they want a single instance with the specified resources.
This is done using the fleet capacity unit 'units'.
We can instead specify the minimum resources we want and set the minimum resources of the instance above this so that a single instance fulfills the reservation.
This means we would specify the fleet capacity unit as 'memory-mib'.
This difference is important as we cannot change the capacity unit with forge modify, so to scale up/down resources we can only change the total capacity.
For master and single instances, setting the capacity unit to 'units' means we effectively lose the ability to modify the fleet as we do not want more units, but more resources.
For workers this is less of an issue but we should prefer 'memory-mib'.
This commit sets it so that single and master instances use 'memory-mib' unless otherwise specified.