Update Forge multi-architecture handling
The way that multi-architecture support was
added in v1.4.0rc1 did not adequately
construct the fleet request. After
discussing with AWS Support, the recommended
configuration has been implemented in this
update. In testing so far this has still
not worked as expected. There is a support
case open with AWS regarding this, but for
now Forge implements what is best practice
to our knowledge.